backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymysql==1.1.0
google-generativeai>=0.5.2
python-dotenv==1.0.0
gunicorn==21.2.0
